:truck: move `internal/version.go` to `internal/app/version.go`
