+++ internal/app/version.go
@@ -1,3 +1,3 @@
-package internal
+package app
 
 const FLOWG_VERSION = "v0.2.0"
